rewrite the examples with single file component
