+++ example/example.html
@@ -5,9 +5,7 @@
     <title></title>
   </head>
   <body>
-    <div id="el">
-      <app prop-message="World!"></app>
-    </div>
+    <div id="app"></div>
     <script src="build.js"></script>
   </body>
 </html>
