+++ example/example.ts
@@ -1,50 +1,10 @@
 import Vue = require('vue')
-import Component from '../lib/index'
-
-@Component({
-  props: {
-    propMessage: String
-  },
-  template: `
-    <div>
-      <input v-model="msg">
-      <p>prop: {{propMessage}}</p>
-      <p>msg: {{msg}}</p>
-      <p>helloMsg: {{helloMsg}}</p>
-      <p>computed msg: {{computedMsg}}</p>
-      <button @click="greet">Greet</button>
-    </div>
-  `
-})
-class App extends Vue {
-  propMessage: string
-
-  // inital data
-  msg: number = 123
-
-  // use prop values for initial data
-  helloMsg: string = 'Hello, ' + this.propMessage
-
-  // lifecycle hook
-  mounted () {
-    this.greet()
-  }
-
-  // computed
-  get computedMsg () {
-    return 'computed ' + this.msg
-  }
-
-  // method
-  greet () {
-    alert('greeting: ' + this.msg)
-  }
-}
+import App from './App.vue'
 
 // mount
 new Vue({
-  el: '#el',
-  components: {
-    App
-  }
+  el: '#app',
+  render: h => h(App, {
+    props: { propMessage: 'World' }
+  })
 })
