+++ example/sfc.d.ts
@@ -1,0 +1,4 @@
+declare module "*.vue" {
+  import Vue = require('vue')
+  export default typeof Vue
+}
