+++ package.json
@@ -43,15 +43,18 @@
     "babel-preset-es2015": "^6.18.0",
     "chai": "^3.5.0",
     "chai-spies": "^0.7.1",
+    "css-loader": "^0.26.1",
     "mocha": "^3.1.2",
     "node-libs-browser": "^1.0.0",
     "rimraf": "^2.5.4",
     "rollup": "^0.37.0",
     "rollup-plugin-replace": "^1.1.1",
-    "ts-loader": "^0.9.5",
+    "ts-loader": "^1.3.3",
     "typescript": "^2.0.6",
     "uglify-js": "^2.7.5",
-    "vue": "^2.1.6",
-    "webpack": "^2.1.0-beta.27"
+    "vue": "^2.1.10",
+    "vue-loader": "^10.0.2",
+    "vue-template-compiler": "^2.1.10",
+    "webpack": "^2.2.0"
   }
 }
